+++ src/main/pluginsServer.ts
@@ -9,8 +9,9 @@ import * as schedule from 'node-schedule'
 import { defaultConfig } from '../config/config'
 import { JobQueueConsumerControl, PluginsServer, PluginsServerConfig, Queue, ScheduleControl } from '../types'
 import { createServer } from '../utils/db/server'
+import { killProcess } from '../utils/kill'
 import { status } from '../utils/status'
-import { createRedis, delay, getPiscinaStats, killProcess } from '../utils/utils'
+import { createRedis, delay, getPiscinaStats } from '../utils/utils'
 import { startQueue } from './ingestion-queues/queue'
 import { startJobQueueConsumer } from './job-queues/job-queue-consumer'
 import { createMmdbServer, performMmdbStalenessCheck, prepareMmdb } from './services/mmdb'
