+++ src/utils/db/server.ts
@@ -14,8 +14,9 @@ import { defaultConfig } from '../../config/config'
 import { JobQueueManager } from '../../main/job-queues/job-queue-manager'
 import { PluginsServer, PluginsServerConfig } from '../../types'
 import { EventsProcessor } from '../../worker/ingestion/process-event'
+import { killProcess } from '../kill'
 import { status } from '../status'
-import { createPostgresPool, createRedis, killProcess, logOrThrowJobQueueError, UUIDT } from '../utils'
+import { createPostgresPool, createRedis, logOrThrowJobQueueError, UUIDT } from '../utils'
 import { DB } from './db'
 import { KafkaProducerWrapper } from './kafka-producer-wrapper'
